refactor(crdt): split the CRDT store bootstrap out of crdt-provider and drop its max-lines override
Move the persistence seam — the CrdtPersistence interface, the preflight/
quarantine/restore dance and the native-binding probe — into a new
crdt-persistence.ts exposing openCrdtPersistence(storagePath). The moved code is
byte-identical apart from returning the persistence instead of assigning
this.persistence, and doInitPersistence keeps the this.persistenceReady flag.

crdt-provider.ts drops from 842 to 748 ESLint-counted lines (cap 800), so its
entry in the large-main-process-module max-lines override is removed. Nothing in
doc identity, refcounting, the LRU pool, compaction or sourceWindowId tagging is
touched.
